refactor: 의도를 드러내는 이름 선택하기
