+++ jongeuni/src/jongeuni/chapther10/RegularPhone.java
@@ -1,0 +1,23 @@
+package jongeuni.chapther10;
+
+import java.time.Duration;
+import jongeuni.chapther02.Money;
+import lombok.Getter;
+
+@Getter
+public class RegularPhone extends AbstractPhone{
+
+    private Money amount;
+    private Duration seconds;
+
+    public RegularPhone(Money amount, Duration seconds) {
+        this.amount = amount;
+        this.seconds = seconds;
+    }
+
+
+    @Override
+    protected Money calculateCallFee(Call call) {
+        return amount.times(call.getDuration().getSeconds() / seconds.getSeconds());
+    }
+}
